Reuse the add view from the categories screen
