+++ WooCommerce/src/main/res/layout/fragment_product_categories_list.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <com.woocommerce.android.ui.products.categories.AddProductCategoryView
+        <com.woocommerce.android.ui.products.AddProductElementView
             android:id="@+id/addProductCategoryView"
             style="@style/Woo.Card"
             android:layout_width="match_parent"
@@ -20,6 +20,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
+            app:buttonText="@string/product_add_category"
             tools:visibility="visible" />
 
         <androidx.recyclerview.widget.RecyclerView
