+++ WooCommerce/src/main/res/values/attrs.xml
@@ -153,4 +153,8 @@
         allow for overriding defStyleRes for some stupid reason.
     -->
     <attr name="wcToggleOutlinedButtonStyle" format="reference"/>
+
+    <declare-styleable name="AddProductElementView">
+        <attr name="buttonText" format="string"/>
+    </declare-styleable>
 </resources>
